added orders component

+++ src/main/java/com/nderitu/ecommerce/entity/Coupon.java
@@ -3,6 +3,7 @@ package com.nderitu.ecommerce.entity;
 import jakarta.persistence.*;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 @Entity
@@ -18,7 +19,7 @@ public class Coupon {
 
     private String code;
 
-    private Long discount; //it will be in %
+    private BigDecimal discount; //it will be in %
 
     private Date expirationDate;
 }
